+++ pysat/instruments/pysat_testing.py
@@ -45,7 +45,8 @@ preprocess = mm_test.preprocess
 
 def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
          sim_multi_file_left=False, root_date=None, malformed_index=False,
-         start_time=None, num_samples=86400, test_load_kwarg=None):
+         start_time=None, num_samples=86400, test_load_kwarg=None,
+         max_latitude=90.):
     """Load the test files.
 
     Parameters
@@ -57,16 +58,16 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
         certain instrument aspects for testing)
     inst_id : str or NoneType
         Instrument satellite ID (accepts '')
-    sim_multi_file_right : boolean
+    sim_multi_file_right : bool
         Adjusts date range to be 12 hours in the future or twelve hours beyond
         root_date (default=False)
-    sim_multi_file_left : boolean
+    sim_multi_file_left : bool
         Adjusts date range to be 12 hours in the past or twelve hours before
         root_date (default=False)
     root_date : NoneType
         Optional central date, uses _test_dates if not specified.
         (default=None)
-    malformed_index : boolean
+    malformed_index : bool
         If True, time index will be non-unique and non-monotonic (default=False)
     start_time : dt.timedelta or NoneType
         Offset time of start time since midnight UT. If None, instrument data
@@ -77,6 +78,9 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
         current day. (default=86400)
     test_load_kwarg : any or NoneType
         Testing keyword (default=None)
+    max_latitude : float
+        Latitude simulated as `max_latitude` * cos(theta(t))`, where
+        theta is a linear periodic signal bounded by [0, 2 * pi) (default=90.).
 
     Returns
     -------
@@ -132,7 +136,7 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
     angle = mm_test.generate_fake_data(time_delta.total_seconds(),
                                        uts, period=iperiod['angle'],
                                        data_range=drange['angle'])
-    data['latitude'] = 90.0 * np.cos(angle)
+    data['latitude'] = max_latitude * np.cos(angle)
 
     # Create constant altitude at 400 km
     alt0 = 400.0